// src/findjobnu-api/models/SkillProficiency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FindjobnuService
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SkillProficiency = {
    NUMBER_0: 0,
    NUMBER_1: 1,
    NUMBER_2: 2,
    NUMBER_3: 3
} as const;
export type SkillProficiency = typeof SkillProficiency[keyof typeof SkillProficiency];


export function instanceOfSkillProficiency(value: any): boolean {
    for (const key in SkillProficiency) {
        if (Object.prototype.hasOwnProperty.call(SkillProficiency, key)) {
            if (SkillProficiency[key as keyof typeof SkillProficiency] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SkillProficiencyFromJSON(json: any): SkillProficiency {
    return SkillProficiencyFromJSONTyped(json, false);
}

export function SkillProficiencyFromJSONTyped(json: any, ignoreDiscriminator: boolean): SkillProficiency {
    return json as SkillProficiency;
}

export function SkillProficiencyToJSON(value?: SkillProficiency | null): any {
    return value as any;
}

export function SkillProficiencyToJSONTyped(value: any, ignoreDiscriminator: boolean): SkillProficiency {
    return value as SkillProficiency;
}